move phone api to audio/service.js
