+++ bigbluebutton-html5/imports/api/verto/index.js
@@ -1,7 +1,7 @@
-import {getInStorage} from '/imports/ui/components/app/service';
+import { getInStorage } from '/imports/ui/components/app/service';
 import Users from '/imports/api/users';
 import Auth from '/imports/ui/services/auth';
-import { getVoiceBridge } from '/imports/api/phone';
+import { getVoiceBridge } from '/imports/ui/components/audio/service';
 
 function createVertoUserName() {
   const uid = Auth.userID;
