+++ bigbluebutton-html5/imports/ui/components/actions-bar/service.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import AuthSingleton from '/imports/ui/services/auth/index.js';
 import Users from '/imports/api/users';
-import { joinListenOnly } from '/imports/api/phone';
-import { showModal } from '/imports/ui/components/app/service';
-import { exitAudio } from '/imports/api/phone';
-import Audio from '/imports/ui/components/audio/audio-modal/component';
 
 let isUserPresenter = () => {
 
@@ -18,17 +14,6 @@ let isUserPresenter = () => {
   };
 };
 
-const handleExitAudio = () => {
-  return exitAudio();
-}
-
-const handleJoinAudio = () => {
-  const handleJoinListenOnly = () => joinListenOnly();
-  return showModal(<Audio handleJoinListenOnly={handleJoinListenOnly} />);
-}
-
 export default {
   isUserPresenter,
-  handleJoinAudio,
-  handleExitAudio,
 };
