+++ bigbluebutton-html5/imports/ui/components/audio/enter-audio/container.jsx
@@ -1,6 +1,6 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import { createContainer } from 'meteor/react-meteor-data';
-import {joinListenOnly, joinMicrophone} from '/imports/api/phone';
+import { joinListenOnly, joinMicrophone } from '../service';
 import { clearModal } from '/imports/ui/components/app/service';
 import EnterAudio from './component';
